move default schema code to new file; fix missing param in Reactor
