+++ saltproc/app.py
@@ -6,13 +6,13 @@ import argparse
 import numpy as np
 import json
 import jsonschema
-from jsonschema import Draft202012Validator, validators
 import gc
 import networkx as nx
 import pydotplus
 
 from saltproc import SerpentDepcode, OpenMCDepcode, Simulation, Reactor
 from saltproc import Process, Sparger, Separator, Materialflow
+from ._schema_default import DefaultValidatingValidator
 
 
 def run():
@@ -152,17 +152,18 @@ def read_main_input(main_inp_file):
 
     input_schema = (Path(__file__).parents[0] / 'input_schema.json')
     with open(main_inp_file) as f:
-        j = json.load(f)
-        DefaultValidatingValidator = extend_with_default(Draft202012Validator)
+        obj = json.load(f)
         with open(input_schema) as s:
-            v = json.load(s)
+            schema = json.load(s)
             try:
-                DefaultValidatingValidator(v).validate(j)
+                DefaultValidatingValidator(schema).validate(obj)
                 #jsonschema.validate(instance=j, schema=v)
             except jsonschema.exceptions.ValidationError:
                 print("Your input file is improperly structured.\
                       Please see saltproc/tests/test.json for an example.")
 
+        j = obj
+        print(j['reactor']['timestep_type'])
         # Global input path
         input_path = (Path.cwd() / Path(f.name).parents[0])
 
@@ -219,21 +220,6 @@ def read_main_input(main_inp_file):
 
         return input_path, process_file, dot_file, (
             depcode_input, simulation_input, reactor_input)
-
-def extend_with_default(validator_class):
-    validate_properties = validator_class.VALIDATORS["properties"]
-    def set_defaults(validator, properties, instance, schema):
-        for property, subschema in properties.items():
-            if "default" in subschema:
-                instance.setdefault(property, subschema["default"])
-        for error in validate_properties(
-            validator, properties, instance, schema,
-        ):
-            yield error
-    return validators.extend(
-        validator_class, {"properties" : set_defaults},
-    )
-
 
 def _print_simulation_input_info(simulation_input, depcode_input):
     """Helper function for `run()` """
